Batch: Tidy up.
