--- internal/service/batch/consts.go
@@ -1,5 +1,0 @@
-package batch
-
-const (
-	jobDefinitionStatusInactive = "INACTIVE"
-)
