--- internal/service/batch/status.go
@@ -1,26 +1,0 @@
-package batch
-
-import (
-	"context"
-
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/batch"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-provider-aws/internal/tfresource"
-)
-
-func statusComputeEnvironment(ctx context.Context, conn *batch.Batch, name string) resource.StateRefreshFunc {
-	return func() (interface{}, string, error) {
-		computeEnvironmentDetail, err := FindComputeEnvironmentDetailByName(ctx, conn, name)
-
-		if tfresource.NotFound(err) {
-			return nil, "", nil
-		}
-
-		if err != nil {
-			return nil, "", err
-		}
-
-		return computeEnvironmentDetail, aws.StringValue(computeEnvironmentDetail.Status), nil
-	}
-}
